Move serve command into its own class
